fixed login bug & post bug

+++ client/src/services/comments.js
@@ -1,26 +1,26 @@
 import api from './api-helper';
 
-export const createComment = async (data) => {
-  const resp = await api.post('/comments', { comment: data });
+export const createComment = async (user_id, post_id, data) => {
+  const resp = await api.post(`/users/${user_id}/posts/${post_id}/comments`, { comment: data });
   return resp.data;
 }
 
-export const readAllComment = async () => {
-  const resp = await api.get('/comments');
+export const readAllComment = async (user_id, post_id) => {
+  const resp = await api.get(`/users/${user_id}/posts/${post_id}/comments`);
   return resp.data;
 }
 
-export const readOneComment = async (id) => {
-  const resp = await api.get(`/comments/${id}`);
+export const readOneComment = async (user_id, post_id, id) => {
+  const resp = await api.get(`/users/${user_id}/posts/${post_id}/comments/${id}`);
   return resp.data
 }
 
-export const updateComment = async (id, data) => {
-  const resp = await api.put(`/comments/${id}`, { comment: data });
+export const updateComment = async (user_id, post_id, id, data) => {
+  const resp = await api.put(`/users/${user_id}/posts/${post_id}/comments/${id}`, { comment: data });
   return resp.data
 }
 
-export const destroyComment = async (id) => {
-  const resp = await api.delete(`/comments/${id}`);
+export const destroyComment = async (user_id, post_id, id) => {
+  const resp = await api.delete(`/users/${user_id}/posts/${post_id}/comments/${id}`);
   return resp;
 }
